pyproject.toml
[project]
name = "rag-agent-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "arxiv>=2.3.1",
    "beautifulsoup4>=4.14.3",
    "duckduckgo-search>=8.1.1",
    "fastapi>=0.123.10",
    "google-search-results>=2.4.2",
    "langchain>=1.1.0",
    "langchain-chroma>=1.0.0",
    "langchain-community>=0.4.1",
    "langchain-experimental>=0.4.1",
    "langchain-ollama>=1.0.0",
    "langchain-openai>=1.1.0",
    "langchain-tavily>=0.2.15",
    "langchain-text-splitters>=1.0.0",
    "langgraph>=1.0.4",
    "pymupdf4llm>=0.2.6",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "python-pptx>=1.0.2",
    "scikit-learn>=1.7.2",
    "streamlit>=1.52.0",
    "tavily-python>=0.7.13",
    "uvicorn>=0.38.0",
    "wikipedia>=1.4.0",
    "wolframalpha>=5.1.3",
]
